lots of debugging from all the code written previously
dma doesn't work (mx initialization doesnt exist and after writing some
myself, it's buggy and doesn't send perfectly plus it's not much
faster). switching to nonblocking with interrupts, but it's just as slow

updateDisplay would hang but it's bc of reading an uninitialized var.

drawtext prints to the same spot since it's based on cursor location and
cursor wasn't being updated. fixed now

bg global var is white (0xFFFF) if not set. now just setting before i
use instead of assuming

year in struct couldn't hold 2000, so changing size

bunch more logical errors, like field being is set is not the same as it
shows, placeholder keeps being reset, value wrapping not working
properly, month names printed wrong, etc.

display for clock face still needs to update automatically. not set up
yet. haven't tested timer, alarm, stopwatch

weekend value in RTC is not set automatically. should do something with
that.

general comments: display updating is very, very slow.  haven't tried
other ideas yet, but dma is not enough (and doesn't fully work anyway).
buttons respond quickly and vars are updated well w/o screen, but
display is just slow to show.

+++ project_L0/Core/Inc/clocks.h
@@ -30,7 +30,7 @@ struct times {
 
 // weekday can be omitted? 0 if not set
 struct dates {
-	uint8_t yr;
+	uint16_t yr;
 	uint8_t month;
 	uint8_t date;
 	uint8_t weekday;		// in hal, week starts monday counting from 1
